backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
langchain==0.1.12
langchain-anthropic==0.1.4
langchain-community==0.0.28
pydantic==2.6.4 